GetAccountsChange and PatchAccountsConfiguration

+++ src/main/java/com/tbsoaresvalkms/oanda/client/account/models/Account.java
@@ -1,5 +1,6 @@
 package com.tbsoaresvalkms.oanda.client.account.models;
 
+import com.tbsoaresvalkms.oanda.client.instrument.models.Currency;
 import com.tbsoaresvalkms.oanda.client.order.models.Order;
 import com.tbsoaresvalkms.oanda.client.position.models.Position;
 import com.tbsoaresvalkms.oanda.client.trade.models.TradeSummary;
@@ -21,7 +22,7 @@ public class Account {
     private BigDecimal resettablePL;
     private LocalDateTime resettabledPLTime;
     private BigDecimal commission;
-    private Double marginRate;
+    private BigDecimal marginRate;
     private LocalDateTime marginCallEnterTime;
     private Integer marginCallExtensionCount;
     private LocalDateTime lastMarginCallExtensionTime;
@@ -37,11 +38,11 @@ public class Account {
     private BigDecimal marginCloseoutUnrealizedPL;
     private BigDecimal marginCloseoutNAV;
     private BigDecimal marginCloseoutMarginUsed;
-    private Double marginCloseoutPercent;
-    private Double marginCloseoutPositionValue;
+    private BigDecimal marginCloseoutPercent;
+    private BigDecimal marginCloseoutPositionValue;
     private BigDecimal withdrawalLimit;
     private BigDecimal marginCallMarginUsed;
-    private Double marginCallPercent;
+    private BigDecimal marginCallPercent;
     private String lastTransactionID;
     private List<TradeSummary> trades;
     private List<Position> positions;
